test: move spawn EINVAL regression test to test/regression/issue/
The gate check requires regression tests to live under test/**, not
inside the package's own test directory. Move the resolveCommand
coverage to test/regression/issue/29636.test.ts and drop the duplicate
inside packages/bun-debug-adapter-protocol.
